requirements/base.txt
asgiref==3.11.1
attrs==26.1.0
Django>=4.2,<6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==69.0.3
sqlparse==0.5.5
uritemplate==4.2.0
wheel==0.46.3
whitenoise==6.12.0